feat: Add pytorch backend module (#35)
* chore: rename.

* wip: change state shape.

* feat: add heuristic kernel parameter selection to xl_chunk_size kernel.

* wip: add arbitrary sequence length wrapper + test. Test fails atm.

* fix: arbitrary sequence length wrapper.

* feat: make state dtype of trition step fused kernel configurable.

* feat: add test for limit + xl chunk size kernels.

* chore: improve doc.

* fix: native sequence return states.

* feat: enable single step in arbitrary sequence length.

* chore: add mlstm sequence kernel registry.

* chore: fix typo.

* feat: add preliminary backend module.

* chore: support different state dtype in triton step kernel. Store outputs only from a single thread block, i.e. avoid duplicate store.

* feat: fully support state dtype in recurrent sequence loop.

* fix: handle the case where target chunk size is smaller than default block size.

* feat: add pad zeros kernel wrapper.

* feat: integrate pad with zeros into backend module.

* chore: add doc to mlstm backend config.

* chore: add doc to forward pass.

* fix: typo. add doc.

* chore: improve error message.

* chore: clean up backend module.

* fix: merge

* chore: change number format.

---------

+++ mlstm_kernels/torch/utils.py
@@ -30,7 +30,10 @@ def contiguous(fn):
         return fn(
             ctx,
             *(i if not isinstance(i, torch.Tensor) else i.contiguous() for i in args),
-            **{k: (v if not isinstance(v, torch.Tensor) else v.contiguous()) for k, v in kwargs.items()},
+            **{
+                k: (v if not isinstance(v, torch.Tensor) else v.contiguous())
+                for k, v in kwargs.items()
+            },
         )
 
     return wrapper
@@ -41,7 +44,10 @@ def contiguous_noctx(fn):
     def wrapper(*args, **kwargs):
         return fn(
             *(i if not isinstance(i, torch.Tensor) else i.contiguous() for i in args),
-            **{k: (v if not isinstance(v, torch.Tensor) else v.contiguous()) for k, v in kwargs.items()},
+            **{
+                k: (v if not isinstance(v, torch.Tensor) else v.contiguous())
+                for k, v in kwargs.items()
+            },
         )
 
     return wrapper
@@ -52,7 +58,7 @@ def torch2triton_dtype(dtype):
 
 
 def to_numpy(tensor: torch.Tensor) -> np.ndarray:
-    return tensor.detach().cpu().numpy()
+    return tensor.detach().cpu().to(dtype=torch.float64).numpy()
 
 
 def tensor_or_none(x):
